TypeError: Type not convertible to expected type.
Type enum ConditionalEscrow.Team is not implicitly convertible to expected type uint8.
        scores[Team.SAN_FRANCISCO_49ERS] > scores[Team.KANSAS_CITY_CHIEFS]

+++ contracts/ConditionalEscrow.sol
@@ -29,7 +29,7 @@ contract ConditionalEscrow is Escrow, WhitelistAdminRole {
     }
     
     function setWinningTeam() public onlyWhitelistAdmin {
-        scores[Team.SAN_FRANCISCO_49ERS] > scores[Team.KANSAS_CITY_CHIEFS]
+        scores[uint8(Team.SAN_FRANCISCO_49ERS)] > scores[uint8(Team.KANSAS_CITY_CHIEFS)]
             ? winningTeam = Team.SAN_FRANCISCO_49ERS
             : winningTeam = Team.KANSAS_CITY_CHIEFS;
         gameEnded = true;
